Add RoutesLocalization filter
This filter translates your routes to a target language using a simple
tree structure specified in your I18n translation file.
The route segments are matched against the tree when generating urls and
matched against a "reversed" tree upon url recognition.

The process might be a bit overkill but it works™. :)